added way to store past transactions automatically

+++ unlonely-alpha/server/src/entities/Channel/channelResolvers.ts
@@ -44,9 +44,9 @@ export const resolvers = {
       { data }: { data: channelService.IDeleteChannelInput },
       ctx: Context
     ) => {
-      // if (!ctx.user || !ctx.userIsAuthed) {
-      //   throw new AuthenticationError("User is not authenticated");
-      // }
+      if (!ctx.user || !ctx.userIsAuthed) {
+        throw new AuthenticationError("User is not authenticated");
+      }
       return channelService.deleteChannel(data, ctx);
     },
     migrateChannelToLivepeer: (
